+++ example/app/bootstrap/bootstrap-example.component.ts
@@ -18,22 +18,19 @@ import {BOOTSTRAP_EXAMPLE_MODEL} from "./bootstrap-example.model";
 export class BootstrapExampleComponent implements OnInit {
 
     dynamicFormModel: DynamicFormModel;
-    dynamicFormService: DynamicFormService;
-
     form: FormGroup;
     
     exampleCheckboxControl: FormControl;
     exampleCheckboxModel: DynamicCheckboxModel;
 
-    constructor(dynamicFormService: DynamicFormService) {
+    constructor(private dynamicFormService: DynamicFormService) {
 
         this.dynamicFormModel = BOOTSTRAP_EXAMPLE_MODEL;
-        this.dynamicFormService = dynamicFormService;
     }
 
     ngOnInit() {
 
-        this.form = this.dynamicFormService.createFormGroup(this.dynamicFormModel);
+        this.form = this.dynamicFormService.createFormGroup(this.dynamicFormModel.group);
 
         this.exampleCheckboxControl = <FormControl> this.form.controls["exampleCheckbox"]; // Type assertion for having updateValue method available
         this.exampleCheckboxModel = <DynamicCheckboxModel> this.dynamicFormModel.findById("exampleCheckbox");
@@ -42,7 +39,7 @@ export class BootstrapExampleComponent implements OnInit {
 
     set modelEdit(value: string) {
         try {
-            this.dynamicFormModel.items = JSON.parse(value);
+            this.dynamicFormModel.group = JSON.parse(value);
 
         } catch (e) {
             // Just do nothing
@@ -50,11 +47,6 @@ export class BootstrapExampleComponent implements OnInit {
     }
 
     get modelEdit() {
-        return JSON.stringify(this.dynamicFormModel.items, null, 2);
-    }
-
-    test() {
-        this.exampleCheckboxModel.disabled = !this.exampleCheckboxModel.disabled;
-        this.dynamicFormModel.items[1].value = "42";
+        return JSON.stringify(this.dynamicFormModel.group, null, 2);
     }
 }