+++ example/app/foundation/foundation-example.component.ts
@@ -22,43 +22,22 @@ import {FOUNDATION_EXAMPLE_MODEL} from "./foundation-example.model";
 export class FoundationExampleComponent implements OnInit {
 
     dynamicFormModel: DynamicFormModel;
-    dynamicFormService: DynamicFormService;
-
     form: FormGroup;
     
     exampleCheckboxControl: FormControl;
     exampleCheckboxModel: DynamicCheckboxModel;
 
-    constructor(dynamicFormService: DynamicFormService) {
+    constructor(private dynamicFormService: DynamicFormService) {
 
         this.dynamicFormModel = FOUNDATION_EXAMPLE_MODEL;
-        this.dynamicFormService = dynamicFormService;
     }
 
     ngOnInit() {
 
-        this.form = this.dynamicFormService.createFormGroup(this.dynamicFormModel);
+        this.form = this.dynamicFormService.createFormGroup(this.dynamicFormModel.group);
 
         this.exampleCheckboxControl = <FormControl> this.form.controls["exampleCheckbox"]; // Type assertion for having updateValue method available
         this.exampleCheckboxModel = <DynamicCheckboxModel> this.dynamicFormModel.findById("exampleCheckbox");
         //this.exampleCheckboxControl.valueChanges.subscribe((value: string) => console.log("example checkbox field changed to: ", value, typeof value));
-    }
-
-    set modelEdit(value: string) {
-        try {
-            this.dynamicFormModel.items = JSON.parse(value);
-
-        } catch (e) {
-            // Just do nothing
-        }
-    }
-
-    get modelEdit() {
-        return JSON.stringify(this.dynamicFormModel.items, null, 2);
-    }
-
-    test() {
-        this.exampleCheckboxModel.disabled = !this.exampleCheckboxModel.disabled;
-        this.dynamicFormModel.items[1].value = "42";
     }
 }