+++ modules/core/src/model/checkbox/dynamic-checkbox-group.model.ts
@@ -1,20 +1,20 @@
-import {DynamicCheckboxModel} from "./dynamic-checkbox.model";
-import {DynamicFormControlModel} from "../dynamic-form-control.model";
-import {IDynamicFormModel} from "../dynamic-form.model";
 import {getValue} from "../../utils";
+import {DynamicFormAbstractControlModel} from "../dynamic-form-abstract-control.model";
+import {IDynamicFormGroupModel} from "../dynamic-form.model";
+import {DynamicCheckboxModel} from "./dynamic-checkbox.model";
 
 export const DYNAMIC_FORM_CONTROL_TYPE_CHECKBOX_GROUP = "CHECKBOX_GROUP";
 
-export class DynamicCheckboxGroupModel extends DynamicFormControlModel<boolean> implements IDynamicFormModel {
+export class DynamicCheckboxGroupModel extends DynamicFormAbstractControlModel implements IDynamicFormGroupModel {
 
-    items: Array<DynamicCheckboxModel>;
+    group: Array<DynamicCheckboxModel>;
     legend: string;
     
     constructor(config: {} = {}, cls?: {}) {
 
         super(config, cls);
 
-        this.items = getValue(config, "items", []);
+        this.group = getValue(config, "group", []);
         this.legend = getValue(config, "legend", null);
         this.type = DYNAMIC_FORM_CONTROL_TYPE_CHECKBOX_GROUP;
     }