+++ modules/core/src/service/dynamic-form.service.ts
@@ -1,7 +1,12 @@
 import {Injectable} from "@angular/core";
-import {FormBuilder, FormControl, FormGroup, Validators} from "@angular/forms";
-import {IDynamicFormModel} from "../model/dynamic-form.model";
-import {DYNAMIC_FORM_CONTROL_TYPE_CHECKBOX_GROUP, DynamicCheckboxGroupModel} from "../model/checkbox/dynamic-checkbox-group.model";
+import {FormBuilder, FormControl, FormGroup, FormArray, Validators} from "@angular/forms";
+import {DynamicFormAbstractControlModel} from "../model/dynamic-form-abstract-control.model";
+import {DynamicFormControlModel} from "../model/dynamic-form-control.model";
+import {DynamicFormArrayModel, DYNAMIC_FORM_CONTROL_TYPE_ARRAY} from "../model/array/dynamic-form-array.model";
+import {
+    DYNAMIC_FORM_CONTROL_TYPE_CHECKBOX_GROUP,
+    DynamicCheckboxGroupModel
+} from "../model/checkbox/dynamic-checkbox-group.model";
 
 @Injectable()
 export class DynamicFormService {
@@ -12,28 +17,75 @@ export class DynamicFormService {
 
         this.formBuilder = formBuilder;
     }
-    
-    createFormGroup(dynamicFormModel: IDynamicFormModel): FormGroup {
+
+    createFormArray(groups: Array<Array<DynamicFormControlModel<any>>>): FormArray {
+
+        let formArray = [];
+
+        groups.forEach(group => formArray.push(this.createFormGroup(group)));
+
+        return this.formBuilder.array(formArray);
+    }
+
+    createFormGroup(group: Array<DynamicFormAbstractControlModel>): FormGroup {
 
         let formGroup = {};
-        
-        dynamicFormModel.items.forEach(controlModel => {
-            
-            if (controlModel.type !== DYNAMIC_FORM_CONTROL_TYPE_CHECKBOX_GROUP) {
+
+        group.forEach(model => {
+
+            if (model.type === DYNAMIC_FORM_CONTROL_TYPE_ARRAY) {
+
+                let arrayModel = <DynamicFormArrayModel> model;
+
+                formGroup[model.id] = this.createFormArray(arrayModel.groups);
+
+            } else if (model.type === DYNAMIC_FORM_CONTROL_TYPE_CHECKBOX_GROUP) {
+
+                let checkboxGroupModel = <DynamicCheckboxGroupModel> model;
+
+                formGroup[model.id] = this.createFormGroup(checkboxGroupModel.group);
+
+            } else {
+
+                let controlModel = <DynamicFormControlModel<any>> model;
 
                 formGroup[controlModel.id] = new FormControl(
-
                     controlModel.value || null,
                     Validators.compose(controlModel.validators),
                     Validators.composeAsync(controlModel.validatorsAsync)
                 );
-                
-            } else {
-                
-                formGroup[controlModel.id] = this.createFormGroup(<DynamicCheckboxGroupModel> controlModel);
             }
         });
 
         return this.formBuilder.group(formGroup);
+    }
+
+
+    createFormArrayGroup(dynamicFormArrayModel: DynamicFormArrayModel): FormGroup {
+
+        return this.createFormGroup(dynamicFormArrayModel.addGroup());
+    }
+
+    addFormArrayGroup(formArray: FormArray, dynamicFormArrayModel: DynamicFormArrayModel): void {
+
+        formArray.push(this.createFormArrayGroup(dynamicFormArrayModel));
+    }
+
+    insertFormArrayGroup(index: number, formArray: FormArray, dynamicFormArrayModel: DynamicFormArrayModel): void {
+
+        formArray.insert(index, this.createFormGroup(dynamicFormArrayModel.insertGroup(index)));
+    }
+
+    removeFormArrayGroup(index: number, formArray: FormArray, dynamicFormArrayModel: DynamicFormArrayModel): void {
+
+        formArray.removeAt(index);
+        dynamicFormArrayModel.removeGroup(index);
+    }
+
+    clearFormArray(formArray: FormArray, dynamicFormArrayModel: DynamicFormArrayModel): void {
+
+        while (formArray.length > 0) {
+            this.removeFormArrayGroup(0, formArray, dynamicFormArrayModel);
+        }
     }
 }