+++ modules/ui-basic/src/dynamic-form-basic.component.ts
@@ -1,4 +1,4 @@
-import {Component} from "@angular/core";
+import {Component, forwardRef, ContentChild, TemplateRef} from "@angular/core";
 import {Input} from "@angular/core";
 import {REACTIVE_FORM_DIRECTIVES, FormGroup} from "@angular/forms";
 import {DynamicFormControlModel, DynamicFormControlComponent} from "@ng2-dynamic-forms/core";
@@ -6,8 +6,8 @@ import {DynamicFormControlModel, DynamicFormControlComponent} from "@ng2-dynamic
 export const DYNAMIC_FORM_UI_BASIC = "BASIC";
 
 @Component({
-    
-    directives: [REACTIVE_FORM_DIRECTIVES],
+
+    directives: [REACTIVE_FORM_DIRECTIVES, forwardRef(() => DynamicFormBasicComponent)],
     moduleId: module.id,
     selector: "dynamic-form-basic-control",
     templateUrl: "./dynamic-form-basic.component.html"
@@ -15,9 +15,11 @@ export const DYNAMIC_FORM_UI_BASIC = "BASIC";
 
 export class DynamicFormBasicComponent extends DynamicFormControlComponent {
 
+    @Input() controlGroup: FormGroup;
     @Input() model: DynamicFormControlModel<any>;
-    @Input() form: FormGroup;
-    
+
+    @ContentChild(TemplateRef) contentTemplate;
+
     type: string = DYNAMIC_FORM_UI_BASIC;
 
     constructor() {