+++ modules/ui-bootstrap/src/dynamic-form-bootstrap.component.ts
@@ -1,5 +1,4 @@
-import {Component} from "@angular/core";
-import {Input} from "@angular/core";
+import {Component, Input, ContentChild, TemplateRef} from "@angular/core";
 import {REACTIVE_FORM_DIRECTIVES, FormGroup} from "@angular/forms";
 import {DynamicFormControlComponent, DynamicFormControlModel} from "@ng2-dynamic-forms/core";
 
@@ -15,8 +14,10 @@ export const DYNAMIC_FORM_UI_BOOTSTRAP = "BOOTSTRAP";
 
 export class DynamicFormBootstrapComponent extends DynamicFormControlComponent {
 
+    @Input() controlGroup: FormGroup;
     @Input() model: DynamicFormControlModel<any>;
-    @Input() form: FormGroup;
+
+    @ContentChild(TemplateRef) contentTemplate;
     
     type: string = DYNAMIC_FORM_UI_BOOTSTRAP;