+++ modules/ui-material/src/dynamic-form-material.component.ts
@@ -1,4 +1,4 @@
-import {Component, Input, ViewChild} from "@angular/core";
+import {Component, Input, ViewChild, ContentChild, TemplateRef} from "@angular/core";
 import {FormGroup, REACTIVE_FORM_DIRECTIVES} from "@angular/forms";
 import {MdUniqueSelectionDispatcher} from "@angular2-material/core";
 import {MdCheckbox} from "@angular2-material/checkbox";
@@ -20,8 +20,10 @@ export const DYNAMIC_FORM_UI_MATERIAL = "MATERIAL";
 
 export class DynamicFormMaterialComponent extends DynamicFormControlComponent {
 
+    @Input() controlGroup: FormGroup;
     @Input() model: DynamicFormControlModel<any>;
-    @Input() form: FormGroup;
+
+    @ContentChild(TemplateRef) contentTemplate;
 
     @ViewChild(MdCheckbox) mdCheckbox: MdCheckbox;
     @ViewChild(MdInput) mdInput: MdInput;