+++ modules/ui-primeng/src/dynamic-form-primeng.component.ts
@@ -1,5 +1,4 @@
-import {Component} from "@angular/core";
-import {Input} from "@angular/core";
+import {Component, Input, ContentChild, TemplateRef} from "@angular/core";
 import {REACTIVE_FORM_DIRECTIVES, FormGroup} from "@angular/forms";
 import {DynamicFormControlComponent, DynamicFormControlModel} from "@ng2-dynamic-forms/core";
 import {Checkbox, Dropdown, InputText, InputTextarea, RadioButton, Spinner} from "primeng/primeng";
@@ -16,8 +15,10 @@ export const DYNAMIC_FORM_UI_PRIME_NG = "PRIME_NG";
 
 export class DynamicFormPrimeNGComponent extends DynamicFormControlComponent {
 
+    @Input() controlGroup: FormGroup;
     @Input() model: DynamicFormControlModel<any>;
-    @Input() form: FormGroup;
+
+    @ContentChild(TemplateRef) contentTemplate;
 
     type: string = DYNAMIC_FORM_UI_PRIME_NG;